Propagate close to async thread
Closing the pcb immediately in lwip thread context can leave the async
object with a dangling reference - there might still be a pending call
on the pcb waiting on the lwip thread lock.  This is a use after free.
The close event must be handled in sequence by the queue to correctly
interlock with the async client.

Further, receiving a fin event does not mean the connections is gone; it
means only that the remote end is done sending.  Clients may still send
data.

Ensure that the close is correctly sequenced.