Initial no-slots implementation
Replace the FreeRTOS queue with a hand-rolled linked list, which permits
the lwip thread to clear invalid events (and do so quickly!).
This eliminates the need for the "close slot" checking and streamlines
the design.

This is the first almost-working version; lots of debug code still needs
to be cleaned up.